+++ src/app/galeria-llavero/galeria-llavero.component.html
@@ -7,99 +7,20 @@
     <div class="col-md-4">
       <div class="card mb-4 box-shadow">
         <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/1.JPG" alt="LLAVERO01" width="50" height="350">
+          <img class="card-img-top" src="/assets/img/1-100/1.JPG" alt="LLAVERO01" width="50" height="350"><!-- "{{this.urlImage}}" -->
         </div>
         <div class="card-body">
           <p class="card-text">--Data.--</p>
           <div class="d-flex justify-content-between align-items-center">
             <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
+              <button type="button" class="btn btn-sm btn-outline-secondary" routerLink="/ver-llavero/{{llavero.$key}}">View</button>
               <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
             </div>
           </div>
         </div>
       </div>
     </div>
-    <div class="col-md-4">
-      <div class="card mb-4 box-shadow">
-        <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/2.JPG" alt="LLAVERO02" width="50" height="350">
-        </div>
-        <div class="card-body">
-          <p class="card-text">--Data.--</p>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-              <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-4">
-      <div class="card mb-4 box-shadow">
-        <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/3.JPG" alt="LLAVERO03" width="50" height="300">
-        </div>
-        <div class="card-body">
-          <p class="card-text">--Data.--</p>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-              <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-4">
-      <div class="card mb-4 box-shadow">
-        <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/4.JPG" alt="LLAVERO03" width="50" height="350">
-        </div>
-        <div class="card-body">
-          <p class="card-text">--Data.--</p>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-              <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-4">
-      <div class="card mb-4 box-shadow">
-        <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/5.JPG" alt="LLAVERO03" width="50" height="350">
-        </div>
-        <div class="card-body">
-          <p class="card-text">--Data.--</p>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-              <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div class="col-md-4">
-      <div class="card mb-4 box-shadow">
-        <div class="img">
-          <img class="card-img-top" src="/assets/img/1-100/6.JPG" alt="LLAVERO03" width="50" height="350">
-        </div>
-        <div class="card-body">
-          <p class="card-text">--Data.--</p>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="btn-group">
-              <button type="button" class="btn btn-sm btn-outline-secondary">View</button>
-              <button type="button" class="btn btn-sm btn-outline-secondary">Edit</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+
   </div>
 </div>
