+++ src/app/galeria-llavero/galeria-llavero.component.ts
@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Llavero } from '../models/llavero';
+
+import { CrudService } from '../services/crud.service';
 
 @Component({
   selector: 'app-galeria-llavero',
@@ -6,10 +9,24 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./galeria-llavero.component.css']
 })
 export class GaleriaLlaveroComponent implements OnInit {
+  Llavero: Llavero[];
 
-  constructor() { }
+  constructor(private crudApi: CrudService,) { }
 
   ngOnInit(): void {
+    this.getData();
+  }
+
+  private getData() {
+    let s = this.crudApi.ObtenerLlaveroList(); 
+    s.snapshotChanges().subscribe(data => { // Using snapshotChanges() method to retrieve list of data along with metadata($key)
+      this.Llavero = [];
+      data.forEach(item => {
+        let a = item.payload.toJSON(); 
+        a['$key'] = item.key;
+        this.Llavero.push(a as Llavero);
+      })
+    })
   }
 
 }
