+++ src/app/models/llavero.spec.ts
@@ -1,0 +1,7 @@
+import { Llavero } from './llavero';
+
+describe('Llavero', () => {
+  it('should create an instance', () => {
+    expect(new Llavero()).toBeTruthy();
+  });
+});
