+++ src/app/models/llavero.ts
@@ -1,0 +1,9 @@
+export class Llavero {
+    $key: string;
+    nombre: string; 
+    material: string;
+    numero: number;
+    comentarios: string;
+    pais: string;
+    img: any;
+}
