+++ src/app/services/crud.service.ts
@@ -1,15 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AngularFireDatabase, AngularFireList, AngularFireObject } from '@angular/fire/database';  // Firebase modules for Database, Data list and Single object
+import { Llavero } from '../models/llavero';
 
-export interface Llavero { 
-  $key: string;
-  nombre: string; 
-  material: string;
-  numero: number;
-  comentarios: string;
-  pais: string;
-  img: any;
-}
 
 @Injectable({
   providedIn: 'root'
